Move config to lib/config
