+++ lib/config.js
@@ -1,16 +1,14 @@
 const _ = require('lodash');
 const path = require('path');
-const fse = require('fs-extra');
 const os = require('os');
 const hbs = require('hbs');
-const clarg = require('clarg');
-const clOpts = clarg().opts;
 
-const appPortRaw = clOpts.port || clOpts.p;
+const appPortRaw = process.env.PORT;
 const appPortParsed = parseInt(appPortRaw, 10);
-const appPort = (appPortRaw == appPortParsed) ? appPortParsed : 3000;
+const appPort = (appPortRaw == appPortParsed) ? appPortParsed : 3000; // eslint-disable-line eqeqeq
+const baseUrl = process.env.BASE_URL || `http://localhost:${appPort}`;
 
-const appNetworkRaw = (clOpts.network || clOpts.n || '').trim();
+const appNetworkRaw = (process.env.NETWORK || '').trim();
 const appNetwork = ['public', 'private'].indexOf(appNetworkRaw) === -1 ? 'local' : appNetworkRaw;
 
 const networks = {
@@ -20,16 +18,15 @@ const networks = {
 };
 
 // fixing for DO setup
-const privateInterface = os.networkInterfaces()['eth1'];
+const privateInterface = os.networkInterfaces().eth1;
 if (privateInterface) {
-  networks.private = _.map(_.find(privateInterface, {family: 'IPv4'}), 'address');
+  networks.private = _.map(_.find(privateInterface, { family: 'IPv4' }), 'address');
 }
-
 
 const appNetworkInterface = networks[appNetwork];
 
 const config = {
-  baseUrl: clOpts.url || clOpts.u || 'http://localhost:' + appPort,
+  baseUrl,
   appPort,
   appNetwork,
   appNetworkInterface,
@@ -43,7 +40,7 @@ const config = {
 
 config.views = {
   engine: hbs,
-  partialsPath: config.paths.views + '/partials'
+  partialsPath: `${config.paths.views}/partials`
 };
 
 // ensure request log file
