+++ lib/ffbinLib.js
@@ -2,6 +2,7 @@ const fs = require('fs');
 const path = require('path');
 const ffbinaries = require('ffbinaries');
 const fmt = require('./fmt');
+
 const ffbinPath = path.resolve('./bin');
 
 /**
@@ -40,7 +41,7 @@ function ensureBinaries(callback) {
     return callback();
   }
   console.log(`[${fmt.date()} ${fmt.time()}] Downloading binaries for ${downloadQueue.join(' and ')}.`);
-  ffbinaries.downloadFiles(downloadQueue, {destination: paths.root}, callback);
+  return ffbinaries.downloadFiles(downloadQueue, {destination: paths.root}, callback);
 }
 
 module.exports = {
