+++ views/layout.html
@@ -9,7 +9,7 @@
     <h1>
       <img src="/public/logo.png" class="icon" />
       <a href="/">
-        VotHub <strong>Media Converter API</strong>
+        <strong>Media Converter</strong> API
       </a>
     </h1>
     <h2>ffmpeg as a service</h2>
