+++ views/partials/nav.html
@@ -3,12 +3,8 @@
     <div class="col-xs-12 col-sm-4 col-sm-offset-4 text-center">
       <a href="/">Start</a>
       <a href="/api">API</a>
-      <!-- <a href="/stats">Stats</a> -->
-      <!-- <a href="/about">About</a> -->
     </div>
 
-    <div class="hidden-xs col-sm-4 text-right">
-      This website doesn't spy on you.
-    </div>
+    <div class="hidden-xs col-sm-4 text-right">&nbsp;</div>
   </div>
 </div>
